backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.4
python-dotenv==1.0.1
fake-useragent==2.0.3
aiofiles==24.1.0
tenacity==9.0.0
tldextract==5.1.3
